refactor: Move stores into `store` module
